+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "./hooks";
+import { useAuth } from "./utils/hooks";
 
 function App() {
 	const isAuth = useAuth();
