+++ src/pages/Auth/Login/Login.tsx
@@ -7,7 +7,7 @@ import { TLoginFormInputs, loginSchema } from "../model";
 import TextField from "../../../components/common/TextField/TextField";
 import Button from "../../../components/common/Button/Button";
 import AuthLayout from "../AuthLayout";
-import { useAuth } from "../../../hooks";
+import { useAuth } from "../../../utils/hooks";
 
 const Login: FunctionComponent = () => {
 	useAuth();
